requirements.txt
# Claim Accountability Ledger - Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database (PostgreSQL support)
psycopg2-binary>=2.9.9  # Sync PostgreSQL driver
asyncpg>=0.29.0         # Async PostgreSQL driver (for high-performance)
# sqlalchemy[asyncio]>=2.0.25  # Optional: if you want ORM
# alembic>=1.13.1              # Optional: for migrations

# Cryptography
cryptography>=41.0.0
pynacl>=1.5.0

# Security
argon2-cffi>=23.1.0     # Password hashing (winner of PHC)
python-multipart>=0.0.6 # Form data parsing (for login)

# Utilities
python-dateutil>=2.8.2
httpx>=0.26.0

# Observability
python-json-logger>=2.0.7  # Structured JSON logging
structlog>=24.1.0          # Better structured logging

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
